requirements.txt
annotated-types==0.7.0
blinker==1.9.0
Bootstrap-Flask==2.4.2
boto3==1.37.30
botocore==1.37.30
click==8.1.8
Deprecated==1.2.18
dominate==2.9.1
Flask==3.1.1
Flask-Bootstrap==3.3.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-APScheduler==1.13.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
limits==4.7.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.34.1
numpy==2.2.4
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
plotly==6.0.1
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
rich==13.9.4
s3transfer==0.11.4
scipy==1.15.2
six==1.17.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
visitor==0.1.3
Werkzeug>=3.1.5
wrapt==1.17.2
WTForms==3.2.1
urllib3>=2.5.0
atomicwrites==1.4.1
gitpython>=3.1.45
yt_dlp>=2025.12.8
keyring==25.7.0
pydub>=0.25.1
requests>=2.32.5
python-dotenv>=1.0.0
pytest-playwright>=0.7.0
playwright>=1.50.0
